refactor: Badge variants 파일 제거 및 매핑 위치 정리
